Add Exp Bar & level

+++ Assets/Scripts/UI_Control/KillText.cs
@@ -6,7 +6,7 @@ public class KillText : MonoBehaviour
     public TMP_Text killText ;
     private void FixedUpdate()
     {
-        killText.text = GameManager.score.ToString();
+        killText.text = ((int)GameManager.score).ToString();
     }
